Move PartialEmoji over to a new namespace to avoid circular imports
